requirements.txt
pyqt6
transformers
accelerate
sentencepiece
protobuf          
python-dotenv
bitsandbytes